+++ crates/node/platform/src/task.rs
@@ -1,4 +1,3 @@
-use crate::LOG_TARGET;
 use lazy_static::lazy_static;
 use moon_config::{TaskCommandArgs, TaskConfig, TasksConfigsMap};
 use moon_logger::{color, debug, warn};
@@ -8,6 +7,8 @@ use moon_utils::regex::{UNIX_SYSTEM_COMMAND, WINDOWS_SYSTEM_COMMAND};
 use moon_utils::{process, regex, string_vec};
 use rustc_hash::FxHashMap;
 use std::collections::BTreeMap;
+
+const LOG_TARGET: &str = "moon:node-platform:tasks";
 
 pub type ScriptsMap = FxHashMap<String, String>;
